// src/app/models/legajo/leg_grup_inv_sem.ts
export interface Leg_Grup_Inv_Sem {
  nLegLidGrupInvSem: number
  nTipoLegLidGrupInvSemCodigo: number
  nInvCodigo: number
  cPerCodigo: number
  nLegLidGrupInvSemTitulo: string
  nLegLidGrupInvSemArchivo: any
  nLegLidGrupInvSemEstado: Boolean
  cFile: any
  nLegLidGrupInvSemDescription? : string
  nTipoLegLidGrupInvSemDescription?:string
  cPrdDescripcion?:string
  extension? : string
  cArchivo : string
}